requirements.txt
groq
pydantic
python-dotenv
rich
pymupdf
sounddevice
scipy

